third_party/blink/web_tests/external/wpt/payment-request/payment-request-canmakepayment-method.https-expected.txt

This is a testharness.js-based test.
FAIL If payment method identifier are supported, resolve promise with true. promise_test: Unhandled rejection with value: object "UnknownError: Request failed"
FAIL If request.[[state]] is "created", then return a promise that resolves to true for known method. assert_equals: if it throws, then it must be a NotAllowedError. expected "NotAllowedError" but got "UnknownError"
FAIL All methods are unsupported promise_test: Unhandled rejection with value: object "UnknownError: Request failed"
FAIL Mix of supported and unsupported methods, at least one method is supported. promise_test: Unhandled rejection with value: object "UnknownError: Request failed"
Harness: the test ran to completion.
